Add TearOut Wrapper component
